separate definitions, loaders and datastructures in stream frontend
