+++ yt/data_objects/tests/test_covering_grid.py
@@ -1,7 +1,7 @@
 import numpy as np
 
 from yt.fields.derived_field import ValidateParameter
-from yt.frontends.stream.data_structures import load_particles
+from yt.frontends.stream.loaders import load_particles
 from yt.loaders import load
 from yt.testing import (
     assert_almost_equal,
