+++ yt/data_objects/tests/test_exclude_functions.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-from yt.frontends.stream.data_structures import load_uniform_grid
+from yt.frontends.stream.loaders import load_uniform_grid
 from yt.testing import assert_equal, fake_random_ds
