+++ yt/data_objects/time_series.py
@@ -11,7 +11,6 @@ from yt.config import ytcfg
 from yt.data_objects.analyzer_objects import AnalysisTask, create_quantity_proxy
 from yt.data_objects.particle_trajectories import ParticleTrajectories
 from yt.funcs import ensure_list, issue_deprecation_warning, iterable, mylog
-from yt.loaders import load
 from yt.units.yt_array import YTArray, YTQuantity
 from yt.utilities.exceptions import YTException, YTOutputNotIdentified
 from yt.utilities.object_registries import (
