+++ yt/frontends/stream/data_structures.py
@@ -2,7 +2,6 @@ import os
 import time
 import uuid
 import weakref
-from collections import defaultdict
 from itertools import chain, product, repeat
 from numbers import Number as numeric_type
 
@@ -15,24 +14,14 @@ from yt.data_objects.particle_unions import ParticleUnion
 from yt.data_objects.static_output import Dataset, ParticleFile
 from yt.data_objects.unions import MeshUnion
 from yt.data_objects.unstructured_mesh import SemiStructuredMesh, UnstructuredMesh
-from yt.frontends.exodus_ii.util import get_num_pseudo_dims
 from yt.frontends.sph.data_structures import SPHParticleIndex
-from yt.funcs import ensure_list, issue_deprecation_warning, iterable
+from yt.funcs import ensure_list
 from yt.geometry.geometry_handler import YTDataChunk
-from yt.geometry.grid_container import GridTree, MatchPointsToGrids
 from yt.geometry.grid_geometry_handler import GridIndex
 from yt.geometry.oct_container import OctreeContainer
 from yt.geometry.oct_geometry_handler import OctreeIndex
 from yt.geometry.unstructured_mesh_handler import UnstructuredIndex
-from yt.units.yt_array import YTQuantity, uconcatenate
-from yt.utilities.decompose import decompose_array, get_psize
-from yt.utilities.exceptions import (
-    YTIllDefinedAMR,
-    YTInconsistentGridFieldShape,
-    YTInconsistentGridFieldShapeGridDims,
-    YTInconsistentParticleFieldShape,
-)
-from yt.utilities.flagging_methods import FlaggingGrid
+from yt.units.yt_array import YTQuantity
 from yt.utilities.io_handler import io_registry
 from yt.utilities.lib.cykdtree import PyKDTree
 from yt.utilities.lib.misc_utilities import get_box_grids_level
@@ -42,6 +31,7 @@ from yt.utilities.lib.particle_kdtree_tools import (
 )
 from yt.utilities.logger import ytLogger as mylog
 
+from .definitions import process_data, set_particle_types
 from .fields import StreamFieldInfo
 
 
@@ -374,749 +364,6 @@ class StreamDictFieldHandler(dict):
         return fields
 
 
-def set_particle_types(data):
-    particle_types = {}
-    for key in data.keys():
-        if key == "number_of_particles":
-            continue
-        if len(data[key].shape) == 1:
-            particle_types[key] = True
-        else:
-            particle_types[key] = False
-    return particle_types
-
-
-def assign_particle_data(ds, pdata, bbox):
-
-    """
-    Assign particle data to the grids using MatchPointsToGrids. This
-    will overwrite any existing particle data, so be careful!
-    """
-
-    for ptype in ds.particle_types_raw:
-        check_fields = [(ptype, "particle_position_x"), (ptype, "particle_position")]
-        if all(f not in pdata for f in check_fields):
-            pdata_ftype = {}
-            for f in [k for k in sorted(pdata)]:
-                if not hasattr(pdata[f], "shape"):
-                    continue
-                if f == "number_of_particles":
-                    continue
-                mylog.debug("Reassigning '%s' to ('%s','%s')", f, ptype, f)
-                pdata_ftype[ptype, f] = pdata.pop(f)
-            pdata_ftype.update(pdata)
-            pdata = pdata_ftype
-
-    # Note: what we need to do here is a bit tricky.  Because occasionally this
-    # gets called before we property handle the field detection, we cannot use
-    # any information about the index.  Fortunately for us, we can generate
-    # most of the GridTree utilizing information we already have from the
-    # stream handler.
-
-    if len(ds.stream_handler.fields) > 1:
-        pdata.pop("number_of_particles", None)
-        num_grids = len(ds.stream_handler.fields)
-        parent_ids = ds.stream_handler.parent_ids
-        num_children = np.zeros(num_grids, dtype="int64")
-        # We're going to do this the slow way
-        mask = np.empty(num_grids, dtype="bool")
-        for i in range(num_grids):
-            np.equal(parent_ids, i, mask)
-            num_children[i] = mask.sum()
-        levels = ds.stream_handler.levels.astype("int64").ravel()
-        grid_tree = GridTree(
-            num_grids,
-            ds.stream_handler.left_edges,
-            ds.stream_handler.right_edges,
-            ds.stream_handler.dimensions,
-            ds.stream_handler.parent_ids,
-            levels,
-            num_children,
-        )
-
-        grid_pdata = []
-        for _ in range(num_grids):
-            grid = {"number_of_particles": 0}
-            grid_pdata.append(grid)
-
-        for ptype in ds.particle_types_raw:
-            if (ptype, "particle_position_x") in pdata:
-                x, y, z = (pdata[ptype, f"particle_position_{ax}"] for ax in "xyz")
-            elif (ptype, "particle_position") in pdata:
-                x, y, z = pdata[ptype, "particle_position"].T
-            else:
-                raise KeyError(
-                    "Cannot decompose particle data without position fields!"
-                )
-            pts = MatchPointsToGrids(grid_tree, len(x), x, y, z)
-            particle_grid_inds = pts.find_points_in_tree()
-            (assigned_particles,) = (particle_grid_inds >= 0).nonzero()
-            num_particles = particle_grid_inds.size
-            num_unassigned = num_particles - assigned_particles.size
-            if num_unassigned > 0:
-                eps = np.finfo(x.dtype).eps
-                s = np.array(
-                    [
-                        [x.min() - eps, x.max() + eps],
-                        [y.min() - eps, y.max() + eps],
-                        [z.min() - eps, z.max() + eps],
-                    ]
-                )
-                sug_bbox = [
-                    [min(bbox[0, 0], s[0, 0]), max(bbox[0, 1], s[0, 1])],
-                    [min(bbox[1, 0], s[1, 0]), max(bbox[1, 1], s[1, 1])],
-                    [min(bbox[2, 0], s[2, 0]), max(bbox[2, 1], s[2, 1])],
-                ]
-                mylog.warning(
-                    "Discarding %s particles (out of %s) that are outside "
-                    "bounding box. Set bbox=%s to avoid this in the future.",
-                    num_unassigned,
-                    num_particles,
-                    sug_bbox,
-                )
-                particle_grid_inds = particle_grid_inds[assigned_particles]
-                x = x[assigned_particles]
-                y = y[assigned_particles]
-                z = z[assigned_particles]
-            idxs = np.argsort(particle_grid_inds)
-            particle_grid_count = np.bincount(
-                particle_grid_inds.astype("intp"), minlength=num_grids
-            )
-            particle_indices = np.zeros(num_grids + 1, dtype="int64")
-            if num_grids > 1:
-                np.add.accumulate(
-                    particle_grid_count.squeeze(), out=particle_indices[1:]
-                )
-            else:
-                particle_indices[1] = particle_grid_count.squeeze()
-            for i, pcount in enumerate(particle_grid_count):
-                grid_pdata[i]["number_of_particles"] += pcount
-                start = particle_indices[i]
-                end = particle_indices[i + 1]
-                for key in pdata.keys():
-                    if key[0] == ptype:
-                        grid_pdata[i][key] = pdata[key][idxs][start:end]
-
-    else:
-        grid_pdata = [pdata]
-
-    for pd, gi in zip(grid_pdata, sorted(ds.stream_handler.fields)):
-        ds.stream_handler.fields[gi].update(pd)
-        ds.stream_handler.particle_types.update(set_particle_types(pd))
-        npart = ds.stream_handler.fields[gi].pop("number_of_particles", 0)
-        ds.stream_handler.particle_count[gi] = npart
-
-
-def process_data(data, grid_dims=None):
-    new_data, field_units = {}, {}
-    for field, val in data.items():
-        # val is a data array
-        if isinstance(val, np.ndarray):
-            # val is a YTArray
-            if hasattr(val, "units"):
-                field_units[field] = val.units
-                new_data[field] = val.copy().d
-            # val is a numpy array
-            else:
-                field_units[field] = ""
-                new_data[field] = val.copy()
-
-        # val is a tuple of (data, units)
-        elif isinstance(val, tuple) and len(val) == 2:
-            try:
-                assert isinstance(field, (str, tuple)), "Field name is not a string!"
-                assert isinstance(val[0], np.ndarray), "Field data is not an ndarray!"
-                assert isinstance(val[1], str), "Unit specification is not a string!"
-                field_units[field] = val[1]
-                new_data[field] = val[0]
-            except AssertionError as e:
-                raise RuntimeError("The data dict appears to be invalid.\n" + str(e))
-
-        # val is a list of data to be turned into an array
-        elif iterable(val):
-            field_units[field] = ""
-            new_data[field] = np.asarray(val)
-
-        else:
-            raise RuntimeError(
-                "The data dict appears to be invalid. "
-                "The data dictionary must map from field "
-                "names to (numpy array, unit spec) tuples. "
-            )
-
-    data = new_data
-
-    # At this point, we have arrays for all our fields
-    new_data = {}
-    for field in data:
-        n_shape = len(data[field].shape)
-        if isinstance(field, tuple):
-            new_field = field
-        elif n_shape in (1, 2):
-            new_field = ("io", field)
-        elif n_shape == 3:
-            new_field = ("stream", field)
-        else:
-            raise RuntimeError
-        new_data[new_field] = data[field]
-        field_units[new_field] = field_units.pop(field)
-        known_fields = (
-            StreamFieldInfo.known_particle_fields + StreamFieldInfo.known_other_fields
-        )
-        # We do not want to override any of the known ones, if it's not
-        # overridden here.
-        if (
-            any(f[0] == new_field[1] for f in known_fields)
-            and field_units[new_field] == ""
-        ):
-            field_units.pop(new_field)
-    data = new_data
-    # Sanity checking that all fields have the same dimensions.
-    g_shapes = []
-    p_shapes = defaultdict(list)
-    for field in data:
-        f_shape = data[field].shape
-        n_shape = len(f_shape)
-        if n_shape in (1, 2):
-            p_shapes[field[0]].append((field[1], f_shape[0]))
-        elif n_shape == 3:
-            g_shapes.append((field, f_shape))
-    if len(g_shapes) > 0:
-        g_s = np.array([s[1] for s in g_shapes])
-        if not np.all(g_s == g_s[0]):
-            raise YTInconsistentGridFieldShape(g_shapes)
-        if grid_dims is not None:
-            if not np.all(g_s == grid_dims):
-                raise YTInconsistentGridFieldShapeGridDims(g_shapes, grid_dims)
-    if len(p_shapes) > 0:
-        for ptype, p_shape in p_shapes.items():
-            p_s = np.array([s[1] for s in p_shape])
-            if not np.all(p_s == p_s[0]):
-                raise YTInconsistentParticleFieldShape(ptype, p_shape)
-    # Now that we know the particle fields are consistent, determine the number
-    # of particles.
-    if len(p_shapes) > 0:
-        number_of_particles = np.sum([s[0][1] for s in p_shapes.values()])
-    else:
-        number_of_particles = 0
-    return field_units, data, number_of_particles
-
-
-def load_uniform_grid(
-    data,
-    domain_dimensions,
-    length_unit=None,
-    bbox=None,
-    nprocs=1,
-    sim_time=0.0,
-    mass_unit=None,
-    time_unit=None,
-    velocity_unit=None,
-    magnetic_unit=None,
-    periodicity=(True, True, True),
-    geometry="cartesian",
-    unit_system="cgs",
-):
-    r"""Load a uniform grid of data into yt as a
-    :class:`~yt.frontends.stream.data_structures.StreamHandler`.
-
-    This should allow a uniform grid of data to be loaded directly into yt and
-    analyzed as would any others.  This comes with several caveats:
-
-    * Units will be incorrect unless the unit system is explicitly
-      specified.
-    * Some functions may behave oddly, and parallelism will be
-      disappointing or non-existent in most cases.
-    * Particles may be difficult to integrate.
-
-    Particle fields are detected as one-dimensional fields.
-
-    Parameters
-    ----------
-    data : dict
-        This is a dict of numpy arrays or (numpy array, unit spec) tuples.
-        The keys are the field names.
-    domain_dimensions : array_like
-        This is the domain dimensions of the grid
-    length_unit : string
-        Unit to use for lengths.  Defaults to unitless.
-    bbox : array_like (xdim:zdim, LE:RE), optional
-        Size of computational domain in units specified by length_unit.
-        Defaults to a cubic unit-length domain.
-    nprocs: integer, optional
-        If greater than 1, will create this number of subarrays out of data
-    sim_time : float, optional
-        The simulation time in seconds
-    mass_unit : string
-        Unit to use for masses.  Defaults to unitless.
-    time_unit : string
-        Unit to use for times.  Defaults to unitless.
-    velocity_unit : string
-        Unit to use for velocities.  Defaults to unitless.
-    magnetic_unit : string
-        Unit to use for magnetic fields. Defaults to unitless.
-    periodicity : tuple of booleans
-        Determines whether the data will be treated as periodic along
-        each axis
-    geometry : string or tuple
-        "cartesian", "cylindrical", "polar", "spherical", "geographic" or
-        "spectral_cube".  Optionally, a tuple can be provided to specify the
-        axis ordering -- for instance, to specify that the axis ordering should
-        be z, x, y, this would be: ("cartesian", ("z", "x", "y")).  The same
-        can be done for other coordinates, for instance:
-        ("spherical", ("theta", "phi", "r")).
-
-    Examples
-    --------
-
-    >>> bbox = np.array([[0., 1.0], [-1.5, 1.5], [1.0, 2.5]])
-    >>> arr = np.random.random((128, 128, 128))
-    >>> data = dict(density=arr)
-    >>> ds = load_uniform_grid(data, arr.shape, length_unit='cm',
-    ...                        bbox=bbox, nprocs=12)
-    >>> dd = ds.all_data()
-    >>> dd['density']
-    YTArray([ 0.87568064,  0.33686453,  0.70467189, ...,  0.70439916,
-            0.97506269,  0.03047113]) g/cm**3
-    """
-
-    domain_dimensions = np.array(domain_dimensions)
-    if bbox is None:
-        bbox = np.array([[0.0, 1.0], [0.0, 1.0], [0.0, 1.0]], "float64")
-    domain_left_edge = np.array(bbox[:, 0], "float64")
-    domain_right_edge = np.array(bbox[:, 1], "float64")
-    grid_levels = np.zeros(nprocs, dtype="int32").reshape((nprocs, 1))
-    # If someone included this throw it away--old API
-    if "number_of_particles" in data:
-        issue_deprecation_warning(
-            "It is no longer necessary to include "
-            "the number of particles in the data "
-            "dict. The number of particles is "
-            "determined from the sizes of the "
-            "particle fields."
-        )
-        data.pop("number_of_particles")
-    # First we fix our field names, apply units to data
-    # and check for consistency of field shapes
-    field_units, data, number_of_particles = process_data(
-        data, grid_dims=tuple(domain_dimensions)
-    )
-
-    sfh = StreamDictFieldHandler()
-
-    if number_of_particles > 0:
-        particle_types = set_particle_types(data)
-        # Used much further below.
-        pdata = {"number_of_particles": number_of_particles}
-        for key in list(data.keys()):
-            if len(data[key].shape) == 1 or key[0] == "io":
-                if not isinstance(key, tuple):
-                    field = ("io", key)
-                    mylog.debug("Reassigning '%s' to '%s'", key, field)
-                else:
-                    field = key
-                sfh._additional_fields += (field,)
-                pdata[field] = data.pop(key)
-    else:
-        particle_types = {}
-
-    if nprocs > 1:
-        temp = {}
-        new_data = {}
-        for key in data.keys():
-            psize = get_psize(np.array(data[key].shape), nprocs)
-            grid_left_edges, grid_right_edges, shapes, slices = decompose_array(
-                data[key].shape, psize, bbox
-            )
-            grid_dimensions = np.array([shape for shape in shapes], dtype="int32")
-            temp[key] = [data[key][slice] for slice in slices]
-        for gid in range(nprocs):
-            new_data[gid] = {}
-            for key in temp.keys():
-                new_data[gid].update({key: temp[key][gid]})
-        sfh.update(new_data)
-        del new_data, temp
-    else:
-        sfh.update({0: data})
-        grid_left_edges = domain_left_edge
-        grid_right_edges = domain_right_edge
-        grid_dimensions = domain_dimensions.reshape(nprocs, 3).astype("int32")
-
-    if length_unit is None:
-        length_unit = "code_length"
-    if mass_unit is None:
-        mass_unit = "code_mass"
-    if time_unit is None:
-        time_unit = "code_time"
-    if velocity_unit is None:
-        velocity_unit = "code_velocity"
-    if magnetic_unit is None:
-        magnetic_unit = "code_magnetic"
-
-    handler = StreamHandler(
-        grid_left_edges,
-        grid_right_edges,
-        grid_dimensions,
-        grid_levels,
-        -np.ones(nprocs, dtype="int64"),
-        np.zeros(nprocs, dtype="int64").reshape(nprocs, 1),  # particle count
-        np.zeros(nprocs).reshape((nprocs, 1)),
-        sfh,
-        field_units,
-        (length_unit, mass_unit, time_unit, velocity_unit, magnetic_unit),
-        particle_types=particle_types,
-        periodicity=periodicity,
-    )
-
-    handler.name = "UniformGridData"
-    handler.domain_left_edge = domain_left_edge
-    handler.domain_right_edge = domain_right_edge
-    handler.refine_by = 2
-    if np.all(domain_dimensions[1:] == 1):
-        dimensionality = 1
-    elif domain_dimensions[2] == 1:
-        dimensionality = 2
-    else:
-        dimensionality = 3
-    handler.dimensionality = dimensionality
-    handler.domain_dimensions = domain_dimensions
-    handler.simulation_time = sim_time
-    handler.cosmology_simulation = 0
-
-    sds = StreamDataset(handler, geometry=geometry, unit_system=unit_system)
-
-    # Now figure out where the particles go
-    if number_of_particles > 0:
-        # This will update the stream handler too
-        assign_particle_data(sds, pdata, bbox)
-
-    return sds
-
-
-def load_amr_grids(
-    grid_data,
-    domain_dimensions,
-    bbox=None,
-    sim_time=0.0,
-    length_unit=None,
-    mass_unit=None,
-    time_unit=None,
-    velocity_unit=None,
-    magnetic_unit=None,
-    periodicity=(True, True, True),
-    geometry="cartesian",
-    refine_by=2,
-    unit_system="cgs",
-):
-    r"""Load a set of grids of data into yt as a
-    :class:`~yt.frontends.stream.data_structures.StreamHandler`.
-    This should allow a sequence of grids of varying resolution of data to be
-    loaded directly into yt and analyzed as would any others.  This comes with
-    several caveats:
-
-    * Units will be incorrect unless the unit system is explicitly specified.
-    * Some functions may behave oddly, and parallelism will be
-      disappointing or non-existent in most cases.
-    * Particles may be difficult to integrate.
-    * No consistency checks are performed on the index
-
-    Parameters
-    ----------
-
-    grid_data : list of dicts
-        This is a list of dicts. Each dict must have entries "left_edge",
-        "right_edge", "dimensions", "level", and then any remaining entries are
-        assumed to be fields. Field entries must map to an NDArray. The grid_data
-        may also include a particle count. If no particle count is supplied, the
-        dataset is understood to contain no particles. The grid_data will be
-        modified in place and can't be assumed to be static.
-    domain_dimensions : array_like
-        This is the domain dimensions of the grid
-    length_unit : string or float
-        Unit to use for lengths.  Defaults to unitless.  If set to be a string, the bbox
-        dimensions are assumed to be in the corresponding units.  If set to a float, the
-        value is a assumed to be the conversion from bbox dimensions to centimeters.
-    mass_unit : string or float
-        Unit to use for masses.  Defaults to unitless.
-    time_unit : string or float
-        Unit to use for times.  Defaults to unitless.
-    velocity_unit : string or float
-        Unit to use for velocities.  Defaults to unitless.
-    magnetic_unit : string or float
-        Unit to use for magnetic fields.  Defaults to unitless.
-    bbox : array_like (xdim:zdim, LE:RE), optional
-        Size of computational domain in units specified by length_unit.
-        Defaults to a cubic unit-length domain.
-    sim_time : float, optional
-        The simulation time in seconds
-    periodicity : tuple of booleans
-        Determines whether the data will be treated as periodic along
-        each axis
-    geometry : string or tuple
-        "cartesian", "cylindrical", "polar", "spherical", "geographic" or
-        "spectral_cube".  Optionally, a tuple can be provided to specify the
-        axis ordering -- for instance, to specify that the axis ordering should
-        be z, x, y, this would be: ("cartesian", ("z", "x", "y")).  The same
-        can be done for other coordinates, for instance:
-        ("spherical", ("theta", "phi", "r")).
-    refine_by : integer or list/array of integers.
-        Specifies the refinement ratio between levels.  Defaults to 2.  This
-        can be an array, in which case it specifies for each dimension.  For
-        instance, this can be used to say that some datasets have refinement of
-        1 in one dimension, indicating that they span the full range in that
-        dimension.
-
-    Examples
-    --------
-
-    >>> grid_data = [
-    ...     dict(left_edge = [0.0, 0.0, 0.0],
-    ...          right_edge = [1.0, 1.0, 1.],
-    ...          level = 0,
-    ...          dimensions = [32, 32, 32],
-    ...          number_of_particles = 0),
-    ...     dict(left_edge = [0.25, 0.25, 0.25],
-    ...          right_edge = [0.75, 0.75, 0.75],
-    ...          level = 1,
-    ...          dimensions = [32, 32, 32],
-    ...          number_of_particles = 0)
-    ... ]
-    ...
-    >>> for g in grid_data:
-    ...     g["density"] = (np.random.random(g["dimensions"])*2**g["level"], "g/cm**3")
-    ...
-    >>> ds = load_amr_grids(grid_data, [32, 32, 32], length_unit=1.0)
-    """
-
-    domain_dimensions = np.array(domain_dimensions)
-    ngrids = len(grid_data)
-    if bbox is None:
-        bbox = np.array([[0.0, 1.0], [0.0, 1.0], [0.0, 1.0]], "float64")
-    domain_left_edge = np.array(bbox[:, 0], "float64")
-    domain_right_edge = np.array(bbox[:, 1], "float64")
-    grid_levels = np.zeros((ngrids, 1), dtype="int32")
-    grid_left_edges = np.zeros((ngrids, 3), dtype="float64")
-    grid_right_edges = np.zeros((ngrids, 3), dtype="float64")
-    grid_dimensions = np.zeros((ngrids, 3), dtype="int32")
-    number_of_particles = np.zeros((ngrids, 1), dtype="int64")
-    parent_ids = np.zeros(ngrids, dtype="int64") - 1
-    sfh = StreamDictFieldHandler()
-    for i, g in enumerate(grid_data):
-        grid_left_edges[i, :] = g.pop("left_edge")
-        grid_right_edges[i, :] = g.pop("right_edge")
-        grid_dimensions[i, :] = g.pop("dimensions")
-        grid_levels[i, :] = g.pop("level")
-        # If someone included this throw it away--old API
-        if "number_of_particles" in g:
-            issue_deprecation_warning(
-                "It is no longer necessary to include "
-                "the number of particles in the data "
-                "dict. The number of particles is "
-                "determined from the sizes of the "
-                "particle fields."
-            )
-            g.pop("number_of_particles")
-        field_units, data, n_particles = process_data(
-            g, grid_dims=tuple(grid_dimensions[i, :])
-        )
-        number_of_particles[i, :] = n_particles
-        sfh[i] = data
-
-    # We now reconstruct our parent ids, so that our particle assignment can
-    # proceed.
-    mask = np.empty(ngrids, dtype="int32")
-    for gi in range(ngrids):
-        get_box_grids_level(
-            grid_left_edges[gi, :],
-            grid_right_edges[gi, :],
-            grid_levels[gi] + 1,
-            grid_left_edges,
-            grid_right_edges,
-            grid_levels,
-            mask,
-        )
-        ids = np.where(mask.astype("bool"))
-        for ci in ids:
-            parent_ids[ci] = gi
-
-    # Check if the grid structure is properly aligned (bug #1295)
-    for lvl in range(grid_levels.min() + 1, grid_levels.max() + 1):
-        idx = grid_levels.flatten() == lvl
-        dims = domain_dimensions * refine_by ** (lvl - 1)
-        for iax, ax in enumerate("xyz"):
-            cell_edges = np.linspace(
-                domain_left_edge[iax], domain_right_edge[iax], dims[iax], endpoint=False
-            )
-            if set(grid_left_edges[idx, iax]) - set(cell_edges):
-                raise YTIllDefinedAMR(lvl, ax)
-
-    if length_unit is None:
-        length_unit = "code_length"
-    if mass_unit is None:
-        mass_unit = "code_mass"
-    if time_unit is None:
-        time_unit = "code_time"
-    if velocity_unit is None:
-        velocity_unit = "code_velocity"
-    if magnetic_unit is None:
-        magnetic_unit = "code_magnetic"
-
-    particle_types = {}
-
-    for grid in sfh.values():
-        particle_types.update(set_particle_types(grid))
-
-    handler = StreamHandler(
-        grid_left_edges,
-        grid_right_edges,
-        grid_dimensions,
-        grid_levels,
-        parent_ids,
-        number_of_particles,
-        np.zeros(ngrids).reshape((ngrids, 1)),
-        sfh,
-        field_units,
-        (length_unit, mass_unit, time_unit, velocity_unit, magnetic_unit),
-        particle_types=particle_types,
-        periodicity=periodicity,
-    )
-
-    handler.name = "AMRGridData"
-    handler.domain_left_edge = domain_left_edge
-    handler.domain_right_edge = domain_right_edge
-    handler.refine_by = refine_by
-    if np.all(domain_dimensions[1:] == 1):
-        dimensionality = 1
-    elif domain_dimensions[2] == 1:
-        dimensionality = 2
-    else:
-        dimensionality = 3
-    handler.dimensionality = dimensionality
-    handler.domain_dimensions = domain_dimensions
-    handler.simulation_time = sim_time
-    handler.cosmology_simulation = 0
-
-    sds = StreamDataset(handler, geometry=geometry, unit_system=unit_system)
-    return sds
-
-
-def refine_amr(base_ds, refinement_criteria, fluid_operators, max_level, callback=None):
-    r"""Given a base dataset, repeatedly apply refinement criteria and
-    fluid operators until a maximum level is reached.
-
-    Parameters
-    ----------
-    base_ds : ~yt.data_objects.static_output.Dataset
-        This is any static output.  It can also be a stream static output, for
-        instance as returned by load_uniform_data.
-    refinement_critera : list of :class:`~yt.utilities.flagging_methods.FlaggingMethod`
-        These criteria will be applied in sequence to identify cells that need
-        to be refined.
-    fluid_operators : list of :class:`~yt.utilities.initial_conditions.FluidOperator`
-        These fluid operators will be applied in sequence to all resulting
-        grids.
-    max_level : int
-        The maximum level to which the data will be refined
-    callback : function, optional
-        A function that will be called at the beginning of each refinement
-        cycle, with the current dataset.
-
-    Examples
-    --------
-    >>> domain_dims = (32, 32, 32)
-    >>> data = np.zeros(domain_dims) + 0.25
-    >>> fo = [ic.CoredSphere(0.05, 0.3, [0.7,0.4,0.75], {"Density": (0.25, 100.0)})]
-    >>> rc = [fm.flagging_method_registry["overdensity"](8.0)]
-    >>> ug = load_uniform_grid({'Density': data}, domain_dims, 1.0)
-    >>> ds = refine_amr(ug, rc, fo, 5)
-    """
-
-    # If we have particle data, set it aside for now
-
-    number_of_particles = np.sum(
-        [grid.NumberOfParticles for grid in base_ds.index.grids]
-    )
-
-    if number_of_particles > 0:
-        pdata = {}
-        for field in base_ds.field_list:
-            if not isinstance(field, tuple):
-                field = ("unknown", field)
-            fi = base_ds._get_field_info(*field)
-            if (
-                fi.sampling_type == "particle"
-                and field[0] in base_ds.particle_types_raw
-            ):
-                pdata[field] = uconcatenate(
-                    [grid[field] for grid in base_ds.index.grids]
-                )
-        pdata["number_of_particles"] = number_of_particles
-
-    last_gc = base_ds.index.num_grids
-    cur_gc = -1
-    ds = base_ds
-    bbox = np.array(
-        [(ds.domain_left_edge[i], ds.domain_right_edge[i]) for i in range(3)]
-    )
-    while ds.index.max_level < max_level and last_gc != cur_gc:
-        mylog.info("Refining another level.  Current max level: %s", ds.index.max_level)
-        last_gc = ds.index.grids.size
-        for m in fluid_operators:
-            m.apply(ds)
-        if callback is not None:
-            callback(ds)
-        grid_data = []
-        for g in ds.index.grids:
-            gd = dict(
-                left_edge=g.LeftEdge,
-                right_edge=g.RightEdge,
-                level=g.Level,
-                dimensions=g.ActiveDimensions,
-            )
-            for field in ds.field_list:
-                if not isinstance(field, tuple):
-                    field = ("unknown", field)
-                fi = ds._get_field_info(*field)
-                if not fi.sampling_type == "particle":
-                    gd[field] = g[field]
-            grid_data.append(gd)
-            if g.Level < ds.index.max_level:
-                continue
-            fg = FlaggingGrid(g, refinement_criteria)
-            nsg = fg.find_subgrids()
-            for sg in nsg:
-                LE = sg.left_index * g.dds + ds.domain_left_edge
-                dims = sg.dimensions * ds.refine_by
-                grid = ds.smoothed_covering_grid(g.Level + 1, LE, dims)
-                gd = dict(
-                    left_edge=LE,
-                    right_edge=grid.right_edge,
-                    level=g.Level + 1,
-                    dimensions=dims,
-                )
-                for field in ds.field_list:
-                    if not isinstance(field, tuple):
-                        field = ("unknown", field)
-                    fi = ds._get_field_info(*field)
-                    if not fi.sampling_type == "particle":
-                        gd[field] = grid[field]
-                grid_data.append(gd)
-
-        ds = load_amr_grids(grid_data, ds.domain_dimensions, bbox=bbox)
-
-        ds.particle_types_raw = base_ds.particle_types_raw
-        ds.particle_types = ds.particle_types_raw
-
-        # Now figure out where the particles go
-        if number_of_particles > 0:
-            # This will update the stream handler too
-            assign_particle_data(ds, pdata, bbox)
-
-        cur_gc = ds.index.num_grids
-
-    return ds
-
-
 class StreamParticleIndex(SPHParticleIndex):
     def __init__(self, ds, dataset_type=None):
         self.stream_handler = ds.stream_handler
@@ -1286,164 +533,6 @@ class StreamParticlesDataset(StreamDataset):
         self.num_neighbors = n_neighbors
 
 
-def load_particles(
-    data,
-    length_unit=None,
-    bbox=None,
-    sim_time=None,
-    mass_unit=None,
-    time_unit=None,
-    velocity_unit=None,
-    magnetic_unit=None,
-    periodicity=(True, True, True),
-    geometry="cartesian",
-    unit_system="cgs",
-    data_source=None,
-):
-    r"""Load a set of particles into yt as a
-    :class:`~yt.frontends.stream.data_structures.StreamParticleHandler`.
-
-    This will allow a collection of particle data to be loaded directly into
-    yt and analyzed as would any others.  This comes with several caveats:
-
-    * There must be sufficient space in memory to contain all the particle
-      data.
-    * Parallelism will be disappointing or non-existent in most cases.
-    * Fluid fields are not supported.
-
-    Note: in order for the dataset to take advantage of SPH functionality,
-    the following two fields must be provided:
-    * ('io', 'density')
-    * ('io', 'smoothing_length')
-
-    Parameters
-    ----------
-    data : dict
-        This is a dict of numpy arrays or (numpy array, unit name) tuples,
-        where the keys are the field names. Particles positions must be named
-        "particle_position_x", "particle_position_y", and "particle_position_z".
-    length_unit : float
-        Conversion factor from simulation length units to centimeters
-    bbox : array_like (xdim:zdim, LE:RE), optional
-        Size of computational domain in units of the length_unit
-    sim_time : float, optional
-        The simulation time in seconds
-    mass_unit : float
-        Conversion factor from simulation mass units to grams
-    time_unit : float
-        Conversion factor from simulation time units to seconds
-    velocity_unit : float
-        Conversion factor from simulation velocity units to cm/s
-    magnetic_unit : float
-        Conversion factor from simulation magnetic units to gauss
-    periodicity : tuple of booleans
-        Determines whether the data will be treated as periodic along
-        each axis
-    data_source : YTSelectionContainer, optional
-        If set, parameters like `bbox`, `sim_time`, and code units are derived
-        from it.
-
-    Examples
-    --------
-
-    >>> pos = [np.random.random(128*128*128) for i in range(3)]
-    >>> data = dict(particle_position_x = pos[0],
-    ...             particle_position_y = pos[1],
-    ...             particle_position_z = pos[2])
-    >>> bbox = np.array([[0., 1.0], [0.0, 1.0], [0.0, 1.0]])
-    >>> ds = load_particles(data, 3.08e24, bbox=bbox)
-
-    """
-
-    domain_dimensions = np.ones(3, "int32")
-    nprocs = 1
-
-    # Parse bounding box
-    if data_source is not None:
-        le, re = data_source.get_bbox()
-        le = le.to_value("code_length")
-        re = re.to_value("code_length")
-        bbox = list(zip(le, re))
-    if bbox is None:
-        bbox = np.array([[0.0, 1.0], [0.0, 1.0], [0.0, 1.0]], "float64")
-    else:
-        bbox = np.array(bbox)
-    domain_left_edge = np.array(bbox[:, 0], "float64")
-    domain_right_edge = np.array(bbox[:, 1], "float64")
-    grid_levels = np.zeros(nprocs, dtype="int32").reshape((nprocs, 1))
-
-    # Parse simulation time
-    if data_source is not None:
-        sim_time = data_source.ds.current_time
-    if sim_time is None:
-        sim_time = 0.0
-    else:
-        sim_time = float(sim_time)
-
-    # Parse units
-    def parse_unit(unit, dimension):
-        if unit is None:
-            unit = "code_" + dimension
-            if data_source is not None:
-                unit = getattr(data_source.ds, dimension + "_unit", unit)
-        return unit
-
-    length_unit = parse_unit(length_unit, "length")
-    mass_unit = parse_unit(mass_unit, "mass")
-    time_unit = parse_unit(time_unit, "time")
-    velocity_unit = parse_unit(velocity_unit, "velocity")
-    magnetic_unit = parse_unit(magnetic_unit, "magnetic")
-
-    # Preprocess data
-    field_units, data, _ = process_data(data)
-    sfh = StreamDictFieldHandler()
-
-    pdata = {}
-    for key in data.keys():
-        if not isinstance(key, tuple):
-            field = ("io", key)
-            mylog.debug("Reassigning '%s' to '%s'", key, field)
-        else:
-            field = key
-        pdata[field] = data[key]
-        sfh._additional_fields += (field,)
-    data = pdata  # Drop reference count
-    particle_types = set_particle_types(data)
-    sfh.update({"stream_file": data})
-    grid_left_edges = domain_left_edge
-    grid_right_edges = domain_right_edge
-    grid_dimensions = domain_dimensions.reshape(nprocs, 3).astype("int32")
-
-    # I'm not sure we need any of this.
-    handler = StreamHandler(
-        grid_left_edges,
-        grid_right_edges,
-        grid_dimensions,
-        grid_levels,
-        -np.ones(nprocs, dtype="int64"),
-        np.zeros(nprocs, dtype="int64").reshape(nprocs, 1),  # Temporary
-        np.zeros(nprocs).reshape((nprocs, 1)),
-        sfh,
-        field_units,
-        (length_unit, mass_unit, time_unit, velocity_unit, magnetic_unit),
-        particle_types=particle_types,
-        periodicity=periodicity,
-    )
-
-    handler.name = "ParticleData"
-    handler.domain_left_edge = domain_left_edge
-    handler.domain_right_edge = domain_right_edge
-    handler.refine_by = 2
-    handler.dimensionality = 3
-    handler.domain_dimensions = domain_dimensions
-    handler.simulation_time = sim_time
-    handler.cosmology_simulation = 0
-
-    sds = StreamParticlesDataset(handler, geometry=geometry, unit_system=unit_system)
-
-    return sds
-
-
 _cis = np.fromiter(
     chain.from_iterable(product([0, 1], [0, 1], [0, 1])), dtype=np.int64, count=8 * 3
 )
@@ -1548,141 +637,6 @@ class StreamHexahedralDataset(StreamDataset):
     _index_class = StreamHexahedralHierarchy
     _field_info_class = StreamFieldInfo
     _dataset_type = "stream_hexahedral"
-
-
-def load_hexahedral_mesh(
-    data,
-    connectivity,
-    coordinates,
-    length_unit=None,
-    bbox=None,
-    sim_time=0.0,
-    mass_unit=None,
-    time_unit=None,
-    velocity_unit=None,
-    magnetic_unit=None,
-    periodicity=(True, True, True),
-    geometry="cartesian",
-    unit_system="cgs",
-):
-    r"""Load a hexahedral mesh of data into yt as a
-    :class:`~yt.frontends.stream.data_structures.StreamHandler`.
-
-    This should allow a semistructured grid of data to be loaded directly into
-    yt and analyzed as would any others.  This comes with several caveats:
-
-    * Units will be incorrect unless the data has already been converted to
-      cgs.
-    * Some functions may behave oddly, and parallelism will be
-      disappointing or non-existent in most cases.
-    * Particles may be difficult to integrate.
-
-    Particle fields are detected as one-dimensional fields. The number of particles
-    is set by the "number_of_particles" key in data.
-
-    Parameters
-    ----------
-    data : dict
-        This is a dict of numpy arrays, where the keys are the field names.
-        There must only be one. Note that the data in the numpy arrays should
-        define the cell-averaged value for of the quantity in in the hexahedral
-        cell.
-    connectivity : array_like
-        This should be of size (N,8) where N is the number of zones.
-    coordinates : array_like
-        This should be of size (M,3) where M is the number of vertices
-        indicated in the connectivity matrix.
-    bbox : array_like (xdim:zdim, LE:RE), optional
-        Size of computational domain in units of the length unit.
-    sim_time : float, optional
-        The simulation time in seconds
-    mass_unit : string
-        Unit to use for masses.  Defaults to unitless.
-    time_unit : string
-        Unit to use for times.  Defaults to unitless.
-    velocity_unit : string
-        Unit to use for velocities.  Defaults to unitless.
-    magnetic_unit : string
-        Unit to use for magnetic fields. Defaults to unitless.
-    periodicity : tuple of booleans
-        Determines whether the data will be treated as periodic along
-        each axis
-    geometry : string or tuple
-        "cartesian", "cylindrical", "polar", "spherical", "geographic" or
-        "spectral_cube".  Optionally, a tuple can be provided to specify the
-        axis ordering -- for instance, to specify that the axis ordering should
-        be z, x, y, this would be: ("cartesian", ("z", "x", "y")).  The same
-        can be done for other coordinates, for instance:
-        ("spherical", ("theta", "phi", "r")).
-
-    """
-
-    domain_dimensions = np.ones(3, "int32") * 2
-    nprocs = 1
-    if bbox is None:
-        bbox = np.array([[0.0, 1.0], [0.0, 1.0], [0.0, 1.0]], "float64")
-    domain_left_edge = np.array(bbox[:, 0], "float64")
-    domain_right_edge = np.array(bbox[:, 1], "float64")
-    grid_levels = np.zeros(nprocs, dtype="int32").reshape((nprocs, 1))
-
-    field_units, data, _ = process_data(data)
-    sfh = StreamDictFieldHandler()
-
-    particle_types = set_particle_types(data)
-
-    sfh.update({"connectivity": connectivity, "coordinates": coordinates, 0: data})
-    # Simple check for axis length correctness
-    if len(data) > 0:
-        fn = list(sorted(data))[0]
-        array_values = data[fn]
-        if array_values.size != connectivity.shape[0]:
-            mylog.error(
-                "Dimensions of array must be one fewer than the coordinate set."
-            )
-            raise RuntimeError
-    grid_left_edges = domain_left_edge
-    grid_right_edges = domain_right_edge
-    grid_dimensions = domain_dimensions.reshape(nprocs, 3).astype("int32")
-
-    if length_unit is None:
-        length_unit = "code_length"
-    if mass_unit is None:
-        mass_unit = "code_mass"
-    if time_unit is None:
-        time_unit = "code_time"
-    if velocity_unit is None:
-        velocity_unit = "code_velocity"
-    if magnetic_unit is None:
-        magnetic_unit = "code_magnetic"
-
-    # I'm not sure we need any of this.
-    handler = StreamHandler(
-        grid_left_edges,
-        grid_right_edges,
-        grid_dimensions,
-        grid_levels,
-        -np.ones(nprocs, dtype="int64"),
-        np.zeros(nprocs, dtype="int64").reshape(nprocs, 1),  # Temporary
-        np.zeros(nprocs).reshape((nprocs, 1)),
-        sfh,
-        field_units,
-        (length_unit, mass_unit, time_unit, velocity_unit, magnetic_unit),
-        particle_types=particle_types,
-        periodicity=periodicity,
-    )
-
-    handler.name = "HexahedralMeshData"
-    handler.domain_left_edge = domain_left_edge
-    handler.domain_right_edge = domain_right_edge
-    handler.refine_by = 2
-    handler.dimensionality = 3
-    handler.domain_dimensions = domain_dimensions
-    handler.simulation_time = sim_time
-    handler.cosmology_simulation = 0
-
-    sds = StreamHexahedralDataset(handler, geometry=geometry, unit_system=unit_system)
-
-    return sds
 
 
 class StreamOctreeSubset(OctreeSubset):
@@ -1796,152 +750,6 @@ class StreamOctreeDataset(StreamDataset):
     _dataset_type = "stream_octree"
 
 
-def load_octree(
-    octree_mask,
-    data,
-    bbox=None,
-    sim_time=0.0,
-    length_unit=None,
-    mass_unit=None,
-    time_unit=None,
-    velocity_unit=None,
-    magnetic_unit=None,
-    periodicity=(True, True, True),
-    over_refine_factor=1,
-    partial_coverage=1,
-    unit_system="cgs",
-):
-    r"""Load an octree mask into yt.
-
-    Octrees can be saved out by calling save_octree on an OctreeContainer.
-    This enables them to be loaded back in.
-
-    This will initialize an Octree of data.  Note that fluid fields will not
-    work yet, or possibly ever.
-
-    Parameters
-    ----------
-    octree_mask : np.ndarray[uint8_t]
-        This is a depth-first refinement mask for an Octree.  It should be
-        of size n_octs * 8 (but see note about the root oct below), where
-        each item is 1 for an oct-cell being refined and 0 for it not being
-        refined.  For over_refine_factors != 1, the children count will
-        still be 8, so there will stil be n_octs * 8 entries. Note that if
-        the root oct is not refined, there will be only one entry
-        for the root, so the size of the mask will be (n_octs - 1)*8 + 1.
-    data : dict
-        A dictionary of 1D arrays.  Note that these must of the size of the
-        number of "False" values in the ``octree_mask``.
-    bbox : array_like (xdim:zdim, LE:RE), optional
-        Size of computational domain in units of length
-    sim_time : float, optional
-        The simulation time in seconds
-    length_unit : string
-        Unit to use for lengths.  Defaults to unitless.
-    mass_unit : string
-        Unit to use for masses.  Defaults to unitless.
-    time_unit : string
-        Unit to use for times.  Defaults to unitless.
-    velocity_unit : string
-        Unit to use for velocities.  Defaults to unitless.
-    magnetic_unit : string
-        Unit to use for magnetic fields. Defaults to unitless.
-    periodicity : tuple of booleans
-        Determines whether the data will be treated as periodic along
-        each axis
-    partial_coverage : boolean
-        Whether or not an oct can be refined cell-by-cell, or whether all
-        8 get refined.
-
-    Example
-    -------
-
-    >>> import yt
-    >>> import numpy as np
-    >>> oct_mask = [8, 0, 0, 0, 0, 8, 0, 8,
-    ...             0, 0, 0, 0, 0, 0, 0, 0,
-    ...             8, 0, 0, 0, 0, 0, 0, 0,
-    ...             0]
-    >>>
-    >>> octree_mask = np.array(oct_mask, dtype=np.uint8)
-    >>> quantities = {}
-    >>> quantities['gas', 'density'] = np.random.random((22, 1), dtype='f8')
-    >>> bbox = np.array([[-10., 10.], [-10., 10.], [-10., 10.]])
-    >>>
-    >>> ds = yt.load_octree(octree_mask=octree_mask,
-    ...                     data=quantities,
-    ...                     bbox=bbox,
-    ...                     over_refine_factor=0,
-    ...                     partial_coverage=0)
-
-    """
-
-    if not isinstance(octree_mask, np.ndarray) or octree_mask.dtype != np.uint8:
-        raise TypeError("octree_mask should be a Numpy array with type uint8")
-
-    nz = 1 << (over_refine_factor)
-    domain_dimensions = np.array([nz, nz, nz])
-    nprocs = 1
-    if bbox is None:
-        bbox = np.array([[0.0, 1.0], [0.0, 1.0], [0.0, 1.0]], "float64")
-    domain_left_edge = np.array(bbox[:, 0], "float64")
-    domain_right_edge = np.array(bbox[:, 1], "float64")
-    grid_levels = np.zeros(nprocs, dtype="int32").reshape((nprocs, 1))
-
-    field_units, data, _ = process_data(data)
-    sfh = StreamDictFieldHandler()
-
-    particle_types = set_particle_types(data)
-
-    sfh.update({0: data})
-    grid_left_edges = domain_left_edge
-    grid_right_edges = domain_right_edge
-    grid_dimensions = domain_dimensions.reshape(nprocs, 3).astype("int32")
-
-    if length_unit is None:
-        length_unit = "code_length"
-    if mass_unit is None:
-        mass_unit = "code_mass"
-    if time_unit is None:
-        time_unit = "code_time"
-    if velocity_unit is None:
-        velocity_unit = "code_velocity"
-    if magnetic_unit is None:
-        magnetic_unit = "code_magnetic"
-
-    # I'm not sure we need any of this.
-    handler = StreamHandler(
-        grid_left_edges,
-        grid_right_edges,
-        grid_dimensions,
-        grid_levels,
-        -np.ones(nprocs, dtype="int64"),
-        np.zeros(nprocs, dtype="int64").reshape(nprocs, 1),  # Temporary
-        np.zeros(nprocs).reshape((nprocs, 1)),
-        sfh,
-        field_units,
-        (length_unit, mass_unit, time_unit, velocity_unit, magnetic_unit),
-        particle_types=particle_types,
-        periodicity=periodicity,
-    )
-
-    handler.name = "OctreeData"
-    handler.domain_left_edge = domain_left_edge
-    handler.domain_right_edge = domain_right_edge
-    handler.refine_by = 2
-    handler.dimensionality = 3
-    handler.domain_dimensions = domain_dimensions
-    handler.simulation_time = sim_time
-    handler.cosmology_simulation = 0
-
-    sds = StreamOctreeDataset(handler, unit_system=unit_system)
-    sds.octree_mask = octree_mask
-    sds.partial_coverage = partial_coverage
-    sds.over_refine_factor = over_refine_factor
-
-    return sds
-
-
 class StreamUnstructuredMesh(UnstructuredMesh):
     _index_offset = 0
 
@@ -1983,250 +791,3 @@ class StreamUnstructuredMeshDataset(StreamDataset):
 
     def _find_particle_types(self):
         pass
-
-
-def load_unstructured_mesh(
-    connectivity,
-    coordinates,
-    node_data=None,
-    elem_data=None,
-    length_unit=None,
-    bbox=None,
-    sim_time=0.0,
-    mass_unit=None,
-    time_unit=None,
-    velocity_unit=None,
-    magnetic_unit=None,
-    periodicity=(False, False, False),
-    geometry="cartesian",
-    unit_system="cgs",
-):
-    r"""Load an unstructured mesh of data into yt as a
-    :class:`~yt.frontends.stream.data_structures.StreamHandler`.
-
-    This should allow an unstructured mesh data to be loaded directly into
-    yt and analyzed as would any others.  Not all functionality for
-    visualization will be present, and some analysis functions may not yet have
-    been implemented.
-
-    Particle fields are detected as one-dimensional fields. The number of
-    particles is set by the "number_of_particles" key in data.
-
-    In the parameter descriptions below, a "vertex" is a 3D point in space, an
-    "element" is a single polyhedron whose location is defined by a set of
-    vertices, and a "mesh" is a set of polyhedral elements, each with the same
-    number of vertices.
-
-    Parameters
-    ----------
-
-    connectivity : list of array_like or array_like
-        This should either be a single 2D array or list of 2D arrays.  If this
-        is a list, each element in the list corresponds to the connectivity
-        information for a distinct mesh. Each array can have different
-        connectivity length and should be of shape (N,M) where N is the number
-        of elements and M is the number of vertices per element.
-    coordinates : array_like
-        The 3D coordinates of mesh vertices. This should be of size (L, D) where
-        L is the number of vertices and D is the number of coordinates per vertex
-        (the spatial dimensions of the dataset). Currently this must be either 2 or 3.
-        When loading more than one mesh, the data for each mesh should be concatenated
-        into a single coordinates array.
-    node_data : dict or list of dicts
-        For a single mesh, a dict mapping field names to 2D numpy arrays,
-        representing data defined at element vertices. For multiple meshes,
-        this must be a list of dicts.  Note that these are not the values as a
-        function of the coordinates, but of the connectivity.  Their shape
-        should be the same as the connectivity.  This means that if the data is
-        in the shape of the coordinates, you may need to reshape them using the
-        `connectivity` array as an index.
-    elem_data : dict or list of dicts
-        For a single mesh, a dict mapping field names to 1D numpy arrays, where
-        each array has a length equal to the number of elements. The data
-        must be defined at the center of each mesh element and there must be
-        only one data value for each element. For multiple meshes, this must be
-        a list of dicts, with one dict for each mesh.
-    bbox : array_like (xdim:zdim, LE:RE), optional
-        Size of computational domain in units of the length unit.
-    sim_time : float, optional
-        The simulation time in seconds
-    mass_unit : string
-        Unit to use for masses.  Defaults to unitless.
-    time_unit : string
-        Unit to use for times.  Defaults to unitless.
-    velocity_unit : string
-        Unit to use for velocities.  Defaults to unitless.
-    magnetic_unit : string
-        Unit to use for magnetic fields. Defaults to unitless.
-    periodicity : tuple of booleans
-        Determines whether the data will be treated as periodic along
-        each axis
-    geometry : string or tuple
-        "cartesian", "cylindrical", "polar", "spherical", "geographic" or
-        "spectral_cube".  Optionally, a tuple can be provided to specify the
-        axis ordering -- for instance, to specify that the axis ordering should
-        be z, x, y, this would be: ("cartesian", ("z", "x", "y")).  The same
-        can be done for other coordinates, for instance:
-        ("spherical", ("theta", "phi", "r")).
-
-    Examples
-    --------
-
-    Load a simple mesh consisting of two tets.
-
-      >>> # Coordinates for vertices of two tetrahedra
-      >>> coordinates = np.array([[0.0, 0.0, 0.5], [0.0, 1.0, 0.5],
-      ...                         [0.5, 1, 0.5], [0.5, 0.5, 0.0],
-      ...                         [0.5, 0.5, 1.0]])
-      >>> # The indices in the coordinates array of mesh vertices.
-      >>> # This mesh has two elements.
-      >>> connectivity = np.array([[0, 1, 2, 4], [0, 1, 2, 3]])
-      >>>
-      >>> # Field data defined at the centers of the two mesh elements.
-      >>> elem_data = {
-      ...     ('connect1', 'elem_field'): np.array([1, 2])
-      ... }
-      >>>
-      >>> # Field data defined at node vertices
-      >>> node_data = {
-      ...     ('connect1', 'node_field'): np.array([[0.0, 1.0, 2.0, 4.0],
-      ...                                           [0.0, 1.0, 2.0, 3.0]])
-      ... }
-      >>>
-      >>> ds = yt.load_unstructured_mesh(connectivity, coordinates,
-      ...                                elem_data=elem_data,
-      ...                                node_data=node_data)
-    """
-
-    dimensionality = coordinates.shape[1]
-    domain_dimensions = np.ones(3, "int32") * 2
-    nprocs = 1
-
-    if elem_data is None and node_data is None:
-        raise RuntimeError("No data supplied in load_unstructured_mesh.")
-
-    if isinstance(connectivity, list):
-        num_meshes = len(connectivity)
-    else:
-        num_meshes = 1
-    connectivity = ensure_list(connectivity)
-
-    if elem_data is None:
-        elem_data = [{} for i in range(num_meshes)]
-    elem_data = ensure_list(elem_data)
-
-    if node_data is None:
-        node_data = [{} for i in range(num_meshes)]
-    node_data = ensure_list(node_data)
-
-    data = [{} for i in range(num_meshes)]
-    for elem_dict, data_dict in zip(elem_data, data):
-        for field, values in elem_dict.items():
-            data_dict[field] = values
-    for node_dict, data_dict in zip(node_data, data):
-        for field, values in node_dict.items():
-            data_dict[field] = values
-    data = ensure_list(data)
-
-    if bbox is None:
-        bbox = [
-            [
-                coordinates[:, i].min() - 0.1 * abs(coordinates[:, i].min()),
-                coordinates[:, i].max() + 0.1 * abs(coordinates[:, i].max()),
-            ]
-            for i in range(dimensionality)
-        ]
-
-    if dimensionality < 3:
-        bbox.append([0.0, 1.0])
-    if dimensionality < 2:
-        bbox.append([0.0, 1.0])
-
-    # handle pseudo-dims here
-    num_pseudo_dims = get_num_pseudo_dims(coordinates)
-    dimensionality -= num_pseudo_dims
-    for i in range(dimensionality, 3):
-        bbox[i][0] = 0.0
-        bbox[i][1] = 1.0
-
-    bbox = np.array(bbox, dtype=np.float64)
-    domain_left_edge = np.array(bbox[:, 0], "float64")
-    domain_right_edge = np.array(bbox[:, 1], "float64")
-    grid_levels = np.zeros(nprocs, dtype="int32").reshape((nprocs, 1))
-
-    field_units = {}
-    particle_types = {}
-    sfh = StreamDictFieldHandler()
-
-    sfh.update({"connectivity": connectivity, "coordinates": coordinates})
-    for i, d in enumerate(data):
-        _f_unit, _data, _ = process_data(d)
-        field_units.update(_f_unit)
-        sfh[i] = _data
-        particle_types.update(set_particle_types(d))
-    # Simple check for axis length correctness
-    if 0 and len(data) > 0:
-        fn = list(sorted(data))[0]
-        array_values = data[fn]
-        if array_values.size != connectivity.shape[0]:
-            mylog.error(
-                "Dimensions of array must be one fewer than the coordinate set."
-            )
-            raise RuntimeError
-    grid_left_edges = domain_left_edge
-    grid_right_edges = domain_right_edge
-    grid_dimensions = domain_dimensions.reshape(nprocs, 3).astype("int32")
-
-    if length_unit is None:
-        length_unit = "code_length"
-    if mass_unit is None:
-        mass_unit = "code_mass"
-    if time_unit is None:
-        time_unit = "code_time"
-    if velocity_unit is None:
-        velocity_unit = "code_velocity"
-    if magnetic_unit is None:
-        magnetic_unit = "code_magnetic"
-
-    # I'm not sure we need any of this.
-    handler = StreamHandler(
-        grid_left_edges,
-        grid_right_edges,
-        grid_dimensions,
-        grid_levels,
-        -np.ones(nprocs, dtype="int64"),
-        np.zeros(nprocs, dtype="int64").reshape(nprocs, 1),  # Temporary
-        np.zeros(nprocs).reshape((nprocs, 1)),
-        sfh,
-        field_units,
-        (length_unit, mass_unit, time_unit, velocity_unit, magnetic_unit),
-        particle_types=particle_types,
-        periodicity=periodicity,
-    )
-
-    handler.name = "UnstructuredMeshData"
-    handler.domain_left_edge = domain_left_edge
-    handler.domain_right_edge = domain_right_edge
-    handler.refine_by = 2
-    handler.dimensionality = dimensionality
-    handler.domain_dimensions = domain_dimensions
-    handler.simulation_time = sim_time
-    handler.cosmology_simulation = 0
-
-    sds = StreamUnstructuredMeshDataset(
-        handler, geometry=geometry, unit_system=unit_system
-    )
-
-    fluid_types = ["all"]
-    for i in range(1, num_meshes + 1):
-        fluid_types += ["connect%d" % i]
-    sds.fluid_types = tuple(fluid_types)
-
-    def flatten(l):
-        return [item for sublist in l for item in sublist]
-
-    sds._node_fields = flatten([[f[1] for f in m] for m in node_data if m])
-    sds._elem_fields = flatten([[f[1] for f in m] for m in elem_data if m])
-    sds.default_field = [f for f in sds.field_list if f[0] == "connect1"][-1]
-
-    return sds
