+++ yt/utilities/tests/test_particle_generator.py
@@ -2,7 +2,8 @@ import numpy as np
 
 import yt.utilities.flagging_methods as fm
 import yt.utilities.initial_conditions as ic
-from yt.loaders import load_uniform_grid, refine_amr
+from yt.frontends.stream.definitions import refine_amr
+from yt.loaders import load_uniform_grid
 from yt.testing import assert_almost_equal, assert_equal
 from yt.units.yt_array import uconcatenate
 from yt.utilities.particle_generator import (
